[Sonnet 4.6] Move Source Tabs outside of Demo
